requirements.txt
Brotli==1.1.0
certifi==2025.10.5
charset-normalizer==3.4.4
filelock==3.20.0
idna==3.11
phonenumbers==9.0.11
psutil==7.1.3
pycryptodomex==3.21.0
pyotp==2.9.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.10.2
PySocks==1.7.1
requests==2.32.5
SQLAlchemy==2.0.44
typing_extensions==4.15.0
urllib3==2.5.0
